% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{affected_reg_effect}
\alias{affected_reg_effect}
\alias{affected_reg_effect.multfsusie}
\title{Find affected region for effect l omics k}
\usage{
affected_reg_effect(affected_reg_effect, ...)

\method{affected_reg_effect}{multfsusie}(multfsusie.obj, l, k, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object}

\item{l}{the effect number}

\item{k}{a digit}

\item{\ldots}{other arguments}
}
\description{
Find affected region for effect l omics k
}
\keyword{internal}
