% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_routine.R
\name{cal_clfsr}
\alias{cal_clfsr}
\alias{cal_clfsr.multfsusie_prior}
\title{Compute conditional local false sign rate}
\usage{
cal_clfsr(G_prior, effect_estimate, list_indx_lst, ...)

\method{cal_clfsr}{multfsusie_prior}(G_prior, effect_estimate, list_indx_lst)
}
\arguments{
\item{G_prior}{multfsusie_prior}

\item{effect_estimate}{output of cal_Bhat_Shat_multfsusie}

\item{list_indx_lst}{list of list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution}
}
\value{
esitmated conditional lfsr
}
\description{
Compute conditional local false sign rate
}
