% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO_mutlfsusie.R
\name{cal_KL_l}
\alias{cal_KL_l}
\alias{cal_KL_l.multfsusie}
\title{Compute KL divergence effect l}
\usage{
cal_KL_l(multfsusie.obj, l, Y, X, D, C, list_indx_lst, ...)

\method{cal_KL_l}{multfsusie}(multfsusie.obj, l, Y, X, list_indx_lst, ...)
}
\arguments{
\item{multfsusie.obj}{a susisF object defined by \code{\link{init_multfsusie_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{Y}{wavelet transformed phenotype}

\item{X}{matrix of covariate}

\item{list_indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
The KL divergence for effect l
}
\description{
Compute KL divergence effect l
}
\keyword{internal}
