% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{cal_partial_resid}
\alias{cal_partial_resid}
\alias{cal_partial_resid.multfsusie}
\title{Compute partial residuals}
\usage{
cal_partial_resid(mvfsusie.obj, l, X, D, indx_lst, ...)

\method{cal_partial_resid}{multfsusie}(
  multfsusie.obj = multfsusie.obj,
  l = l,
  X = X,
  Y = Y,
  list_indx_lst = list_indx_lst
)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object}

\item{\ldots}{other arguments}
}
\description{
Compute partial residuals
}
\keyword{internal}
