% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{check_cs}
\alias{check_cs}
\alias{check_cs.multfsusie}
\title{Check purity credible sets}
\usage{
check_cs(multfsusie.obj, min_purity = 0.5, X, ...)

\method{check_cs}{multfsusie}(multfsusie.obj, min_purity = 0.5, X, allow_empty = TRUE, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by init_multfsusie_obj function}

\item{min_purity}{minimal purity within a CS}

\item{X}{matrix of covariate}

\item{allow_empty}{logical; if TRUE, all unsupported effects may be removed
from the final output.}
}
\value{
a multfsusie.obj without "dummy" credible s
}
\description{
Check purity credible sets
}
\keyword{internal}
