% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{discard_cs}
\alias{discard_cs}
\alias{discard_cs.multfsusie}
\title{Discard credible sets}
\usage{
discard_cs(multfsusie.obj, cs, ...)

\method{discard_cs}{multfsusie}(multfsusie.obj, cs, out_prep = FALSE)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by \code{\link{init_multfsusie_obj}} function}

\item{cs}{vector of integer containing the credible sets to discard}
}
\value{
a multfsusie.obj without "dummy" credible sets
}
\description{
Discard credible sets
}
\keyword{internal}
