% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{get_G_prior}
\alias{get_G_prior}
\alias{get_G_prior.multfsusie}
\title{Acces prior object for multfsusie.obj}
\usage{
get_G_prior(mvfsusie.obj.obj, ...)

\method{get_G_prior}{multfsusie}(multfsusie.obj)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object}
}
\description{
see title
}
\keyword{internal}
