% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{get_lBF}
\alias{get_lBF}
\alias{get_lBF.multfsusie}
\title{Acces log Bayes factors for multfsusie.obj}
\usage{
get_lBF(multfsusie.obj, l, ...)

\method{get_lBF}{multfsusie}(multfsusie.obj, l)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object}

\item{l}{effect to be accesed}
}
\description{
see title
}
\keyword{internal}
