% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_prior.R
\name{get_pi_G_prior.multfsusie_prior}
\alias{get_pi_G_prior.multfsusie_prior}
\title{Extract mixture component from a multfsusie_prior}
\usage{
\method{get_pi_G_prior}{multfsusie_prior}(G_prior)
}
\arguments{
\item{G_prior}{an object of class "multfsusie_prior"}
}
\value{
an object of class "pi_multfsusie" which is a list of two
}
\description{
Extract mixture component from a multfsusie_prior
}
\keyword{internal}
