% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_prior.R
\name{get_pi_G_prior}
\alias{get_pi_G_prior}
\alias{get_pi_G_prior.multfsusie_prior}
\title{Extract mixture component from a multfsusie_prior}
\usage{
get_pi_G_prior(G_prior, ...)

\method{get_pi_G_prior}{multfsusie_prior}(G_prior, ...)
}
\arguments{
\item{G_prior}{an object of class "multfsusie_prior"}

\item{...}{Additional arguments passed to other functions.}
}
\value{
an object of class "pi_multfsusie" which is a list of two elements. The first one
est_pi_u  corresponding of the k ash prior for each of the univariate traits
The second element est_pi_f corresponding of the k  "mixture_normal_per_scale" prior for each of the functional trait
}
\description{
Extract mixture component from a multfsusie_prior
}
\keyword{internal}
