% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{greedy_backfit}
\alias{greedy_backfit}
\alias{greedy_backfit.multfsusie}
\title{Update  susiF via greedy search or backfit}
\usage{
greedy_backfit(multfsusie.obj, verbose, cov_lev, X, min.purity, ...)

\method{greedy_backfit}{multfsusie}(multfsusie.obj, verbose, cov_lev, X, min.purity, ...)
}
\arguments{
\item{multfsusie.obj}{a susiF object defined by \code{\link{init_multfsusie.obj_obj}} function}
}
\value{
multfsusie.obj object
}
\description{
Update  susiF via greedy search or backfit
}
\keyword{internal}
