% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{HMM_regression}
\alias{HMM_regression}
\alias{HMM_regression.multfsusie}
\title{Compute refined estimate using HMM regression}
\usage{
HMM_regression(multfsusie.obj, Y, X, verbose, maxit, ...)

\method{HMM_regression}{multfsusie}(
  multfsusie.obj,
  Y,
  X,
  verbose = TRUE,
  maxit = 5,
  ind_analysis,
  ...
)
}
\arguments{
\item{multfsusie.obj}{a multfsusie.obj object}

\item{Y}{multfsusie.obj data object}

\item{X}{matrix containing the covariates}

\item{verbose}{logical}

\item{maxit}{maximum number of iterations}

\item{...}{Additional arguments passed to other functions.}

\item{ind_analysis}{output of which_notNA_pos}
}
\description{
Compute refined estimate using HMM regression
}
