% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO_mutlfsusie.R
\name{loglik_SFR_post}
\alias{loglik_SFR_post}
\alias{loglik_SFR_post.multfsusie}
\title{Compute posterior expected loglikelihood for  single function regression of effect l}
\usage{
loglik_SFR_post(multfsusie.obj, l, Y, X, ...)

\method{loglik_SFR_post}{multfsusie}(multfsusie.obj, l, Y, X, ind_analysis, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by \code{\link{init_multfsusie_obj}} function}

\item{l}{effect to update}

\item{Y}{wavelet transformed phenotype}

\item{X}{Matrix of covariates}

\item{...}{Additional arguments passed to other functions.}

\item{ind_analysis}{output of which_notNA_pos}
}
\value{
posterior expected loglikelihood for  single function regression of effect l
}
\description{
Compute posterior expected loglikelihood for  single function regression of effect l
}
\keyword{internal}
