% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function.R
\name{mvf_susie_per_level}
\alias{mvf_susie_per_level}
\title{Simulate function under the mash per scale prior}
\usage{
mvf_susie_per_level(
  lev_res = 7,
  n_curve = 3,
  length_grid = 10,
  pi0,
  alpha = 0.8,
  prop_decay = 0.1,
  is.plot = TRUE
)
}
\arguments{
\item{lev_res}{numerical corresponds to the resolution of the simulated function (idealy between 3 and 10)}

\item{n_curve}{dimension of the multivaraite time serie to generate}

\item{length_grid}{vector numerical corresponds to the length of the grid of sigma for mixture component(cf ash)}

\item{pi0}{vector numerical , contain a digit  between 0 and 1, which corresponds to the null proportion ( non assocatied wavelet coefficients)}

\item{alpha}{numeric >0, control smoothness of the curves, should be positive and up 4 in particular d_sl ~  pi_{0,sl}  delta_0 + sum_k  pi_k N(0, 2^{- alpha * s}   sigma_k^2)}

\item{prop_decay}{numeric >0, control the proportion of non zero wavelet coefficient per scale, pi_{0,sl} = 1- exp(-prop_decay*s)}

\item{is.plot}{logical, if true plot the simualted effect}
}
\description{
Simulate function under the mash per scale prior
}
