% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{plot_effect_multfsusiF}
\alias{plot_effect_multfsusiF}
\title{Plot specific effect from susiF object}
\usage{
plot_effect_multfsusiF(
  multfsusie.obj,
  effect = 1,
  lfsr_thresh = 0.05,
  title = "",
  cred.band = TRUE,
  lfsr.curve = TRUE,
  size_line = 2,
  size_point = 4,
  pos_SNP,
  pip_only = FALSE,
  point_shape,
  ...
)
}
\arguments{
\item{multfsusie.obj}{output of the susiF function}

\item{effect}{the index of the effect to be plotted}

\item{lfsr_thresh}{numeric, threshold for the lfsr curve, set to 0.0 by default.}

\item{title}{character}

\item{cred.band}{logical, if TRUE, plot credible bands if
susiF.obj fitted with wavelets regression. Set as TRUE by default.}

\item{lfsr.curve}{logical, if TRUE, plot estimated lfsr of
the effect at each base pair  if susiF.obj fitted with HMM regression.
 Set as TRUE by default.}

\item{size_line}{numeric, width of the plotted lines.}

\item{size_point}{numeric, size of the point.}

\item{pos_SNP}{vector, containing the base pair of the SNPs.}

\item{pip_only}{logical, if TRUE, plot only the pip values. Set as FALSE by default.}

\item{point_shape}{vector, containing the shape of dots.}

\item{\dots}{Other arguments..}
}
\description{
Plot specific effect from susiF object
}
