% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function.R
\name{simu_effect_multfsusie}
\alias{simu_effect_multfsusie}
\title{Simulate effect under the multfsusie model}
\usage{
simu_effect_multfsusie(
  list_lev_res = NULL,
  n_univ = NULL,
  effect_univ,
  output_level = 1
)
}
\arguments{
\item{list_lev_res}{list of level of resoultion (can be set to NULL if non)}

\item{n_univ}{number of univariate trait to be analysed}

\item{effect_univ}{possible effect size of the SNP on the univariate phenotype (if missing the effect size are sample at random as -1 or 1)}

\item{output_level}{if equal to 1 (default) provide simplified output for functionnal effect, setting it to 2 provide more detailled output see
simu_IBSS_per_level in susiF.alpha for additional detail}
}
\description{
Simulate effect under the multfsusie model
}
\examples{


list_lev_res <- list(5,7)
n_univ <- 4
effect <-  simu_effect_multfsusie (list_lev_res=list_lev_res,n_univ=n_univ)

}
