% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{test_stop_cond}
\alias{test_stop_cond}
\alias{test_stop_cond.multfsusie}
\title{Check tolerance for stopping criterion}
\usage{
test_stop_cond(multfsusie.obj, check, cal_obj, Y_data, X, list_indx_lst, ...)

\method{test_stop_cond}{multfsusie}(multfsusie.obj, check, cal_obj, Y, X, list_indx_lst, ...)
}
\description{
Check tolerance for stopping criterion
}
\keyword{internal}
