% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{threshold_set_up}
\alias{threshold_set_up}
\title{Check mark type for multfsusie}
\usage{
threshold_set_up(thresh_u, thresh_f)
}
\arguments{
\item{thresh_u}{vector containing threshold for minimal variance for each univariate trait}

\item{thresh_f}{vector containing threshold for minimal variance for each functional trait}
}
\description{
Allow user to define some threshold valuer for wavelet regression
}
