% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_alpha}
\alias{update_alpha}
\alias{update_alpha.multfsusie}
\alias{update_alpha_hist.multfsusie}
\title{Update alpha   susiF mixture proportion of effect l}
\usage{
update_alpha(multfsusie.obj, l, alpha, ...)

\method{update_alpha}{multfsusie}(multfsusie.obj, l, alpha, ...)

\method{update_alpha_hist}{multfsusie}(multfsusie.obj, discard = FALSE, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie  object defined by \code{\link{init_multfsusie_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{alpha}{vector of p alpha values summing up to one}
}
\value{
multfsusie.obj object
}
\description{
Update alpha   susiF mixture proportion of effect l
}
\keyword{internal}
