% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_alpha_hist}
\alias{update_alpha_hist}
\title{Update alpha_hist   multfsusie object}
\usage{
update_alpha_hist(multfsusie.obj, discard, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by \code{\link{init_multfsusie_obj}} function}

\item{discard}{logical set to FALSE by default, if true remove element of history longer than L}
}
\value{
multfsusie object
}
\description{
Update alpha_hist   multfsusie object
}
\keyword{internal}
