% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_cal_cs}
\alias{update_cal_cs}
\alias{update_cal_cs.multfsusie}
\title{Update multfsusie by computing credible sets}
\usage{
update_cal_cs(multfsusie.obj, cov_lev = 0.95, ...)

\method{update_cal_cs}{multfsusie}(multfsusie.obj, cov_lev = 0.95, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by \code{\link{init_multfsusie_obj}} function}

\item{cov_lev}{numeric between 0 and 1, corresponding to the expected level of coverage of the cs if not specified set to 0.95}
}
\value{
multfsusie object
}
\description{
Update multfsusie by computing credible sets
}
\keyword{internal}
