% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_cal_fit_func}
\alias{update_cal_fit_func}
\alias{update_cal_fit_func.multfsusie}
\title{Update multfsusie by computing posterior curves}
\usage{
update_cal_fit_func(mvfsusie.obj, indx_lst, ...)

\method{update_cal_fit_func}{multfsusie}(multfsusie.obj, list_indx_lst, ...)
}
\arguments{
\item{multfsusie.obj}{a susiF object defined by \code{\link{init_multfsusie_obj}} function}

\item{list_indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
multfsusie object
}
\description{
Update multfsusie by computing posterior curves
}
\keyword{internal}
