% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_cal_fit_u}
\alias{update_cal_fit_u}
\alias{update_cal_fit_u.multfsusie}
\title{Update multfsusie by computing univariate estimates}
\usage{
update_cal_fit_u(multfsusie.obj, ...)

\method{update_cal_fit_u}{multfsusie}(multfsusie.obj, univariate_scale = NULL, ...)
}
\arguments{
\item{multfsusie.obj}{a susiF object defined by init_multfsusie_obj function}

\item{univariate_scale}{scale used to standardize each univariate outcome}
}
\value{
multfsusie object
}
\description{
Update multfsusie by computing univariate estimates
}
\keyword{internal}
