% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_cal_fit_uni}
\alias{update_cal_fit_uni}
\alias{update_cal_fit_uni.multfsusie}
\title{Update multfsusie by computing univariate estimates}
\usage{
update_cal_fit_uni(multfsusie.obj, ...)

\method{update_cal_fit_uni}{multfsusie}(multfsusie.obj, ...)
}
\arguments{
\item{multfsusie.obj}{a susiF object defined by \code{\link{init_multfsusie_obj}} function}
}
\value{
multfsusie object
}
\description{
Update multfsusie by computing univariate estimates
}
\keyword{internal}
