% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_cal_pip}
\alias{update_cal_pip}
\alias{update_cal_pip.multfsusie}
\title{Update multfsusie by computing PiP}
\usage{
update_cal_pip(mvfsusie.obj, ...)

\method{update_cal_pip}{multfsusie}(multfsusie.obj, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by init_multfsusie_obj function}
}
\value{
multfsusie object
}
\description{
Update multfsusie by computing PiP
}
\keyword{internal}
