% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_ELBO}
\alias{update_ELBO}
\alias{update_ELBO.multfsusie}
\title{Update multfsusie log Bayes factor}
\usage{
update_ELBO(multfsusie.obj, ELBO, ...)

\method{update_ELBO}{multfsusie}(multfsusie.obj, ELBO, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by \code{\link{init_multfsusie_obj}} function}

\item{ELBO}{new ELBO value}
}
\value{
multfsusie object
}
\description{
Update multfsusie log Bayes factor
}
\keyword{internal}
