% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_KL}
\alias{update_KL}
\alias{update_KL.multfsusie}
\title{Compute KL divergence effect l
 @param multfsusie.obj a multfsusie object}
\usage{
update_KL(multfsusie.obj, Y, X, list_indx_lst, ...)

\method{update_KL}{multfsusie}(multfsusie.obj, Y, X, list_indx_lst, ind_analysis, ...)
}
\arguments{
\item{Y}{observed response}

\item{X}{matrix of covariates}

\item{list_indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
susiF object
}
\description{
Compute KL divergence effect l
 @param multfsusie.obj a multfsusie object
}
\keyword{internal}
