% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_lfsr_effect}
\alias{update_lfsr_effect}
\alias{update_lfsr_effect.multfsusie}
\title{Update multfsusie.obj lfsr estimates}
\usage{
update_lfsr_effect(multfsusie.obj, ...)

\method{update_lfsr_effect}{multfsusie}(multfsusie.obj, HMM = TRUE, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie.obj object defined by init_multfsusie_obj function}

\item{HMM}{logical if the output is hmm based}
}
\value{
multfsusie.obj object
}
\description{
Update multfsusie.obj lfsr estimates
}
\keyword{internal}
