% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_prior.R
\name{update_prior}
\alias{update_prior}
\alias{update_prior.multfsusie_prior}
\title{Update prior for multfsusie object}
\usage{
update_prior(G_prior, tpi, ...)

\method{update_prior}{multfsusie_prior}(G_prior, tpi, ...)
}
\arguments{
\item{G_prior}{a prior for multfsusie object}

\item{tpi}{mixture proportions}
}
\description{
see title
}
\keyword{internal}
