% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_residual_variance}
\alias{update_residual_variance}
\alias{update_residual_variance.multfsusie}
\title{Update residual variance}
\usage{
update_residual_variance(mvfsusie.obj, sigma2, ...)

\method{update_residual_variance}{multfsusie}(multfsusie.obj, sigma2, ...)
}
\arguments{
\item{sigma2}{the new values for residual variance}

\item{...}{Additional arguments passed to other functions.}

\item{multfsusie.obj}{a multfsusie object}
}
\description{
See title
}
\keyword{internal}
