% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{which_dummy_cs}
\alias{which_dummy_cs}
\alias{which_dummy_cs.multfsusie}
\title{Return which credible sets are  dummy}
\usage{
which_dummy_cs(multfsusie.obj, min_purity = 0.5, X, ...)

\method{which_dummy_cs}{multfsusie}(
  multfsusie.obj,
  min_purity = 0.5,
  X,
  median_crit = FALSE,
  lbf_min,
  allow_empty = FALSE,
  ...
)
}
\arguments{
\item{multfsusie.obj}{a susif object defined by \code{\link{init_susiF_obj}} function}

\item{min_purity}{minimal purity within a CS}

\item{X}{matrix of covariate}

\item{lbf_min}{numeric  discard low purity cs in the IBSS fitting procedure if the largest log Bayes factors is lower than this value}

\item{allow_empty}{logical; if TRUE, a single or final unsupported effect
may be marked as dummy. Keep FALSE during IBSS fitting.}
}
\value{
a list of index corresponding the the dummy effect
}
\description{
Return which credible sets are  dummy
}
\keyword{internal}
