% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO_mutlfsusie.R
\name{Eloglik}
\alias{Eloglik}
\alias{Eloglik.multfsusie}
\title{Expected log likelihood for a   multfsusie   object}
\usage{
Eloglik(multfsusie.obj, Y, X, ind_analysis, ...)

\method{Eloglik}{multfsusie}(multfsusie.obj, Y, X, ind_analysis, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by init_multfsusie_obj function}

\item{Y}{Matrix of outcomes}

\item{X}{Matrix of covariates}

\item{ind_analysis}{list of indices for the analysis}
}
\value{
Expected log likelihood
}
\description{
Expected log likelihood for a   multfsusie   object
}
\keyword{internal}
