% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_routine.R
\name{estimate_residual_variance}
\alias{estimate_residual_variance}
\alias{estimate_residual_variance.multfsusie}
\title{Compute Residual variance}
\usage{
estimate_residual_variance(multfsusie.obj, Y, X, ...)

\method{estimate_residual_variance}{multfsusie}(multfsusie.obj, Y, X, ind_analysis, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object}

\item{Y}{observed response data}

\item{X}{observed covariates}
}
\description{
see title
}
\keyword{internal}
