% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{get_ER2}
\alias{get_ER2}
\alias{get_ER2.multfsusie}
\title{Compute expected sum of squares}
\usage{
get_ER2(mvfsusie.obj, update_DW_tens, X, ...)

\method{get_ER2}{multfsusie}(multfsusie.obj, Y, X)
}
\arguments{
\item{X}{matri of covaraites}

\item{multfsusie.obj}{a multfsusie object}

\item{Y}{matrix of observation}
}
\description{
See title
}
\keyword{internal}
