% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO_mutlfsusie.R
\name{get_objective}
\alias{get_objective}
\alias{get_objective.multfsusie}
\title{Get objective function from data and multfsusie object}
\usage{
get_objective(multfsusie.obj, Y, X, ind_analysis, ...)

\method{get_objective}{multfsusie}(multfsusie.obj, Y, X, ind_analysis, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by init_multfsusie_obj function}

\item{Y}{Matrix of outcomes}

\item{X}{matrix of covariate}

\item{ind_analysis}{output of which_notNA_pos}

\item{...}{Additional arguments passed to other functions.}
}
\value{
objective function value
}
\description{
Get objective function from data and multfsusie object
}
\keyword{internal}
