% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{get_pi}
\alias{get_pi}
\alias{get_pi.multfsusie}
\title{Acces mixture proportion  for multfsusie.obj}
\usage{
get_pi(multfsusie.obj, l, ...)

\method{get_pi}{multfsusie}(multfsusie.obj, l, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object}

\item{l}{effect to be accesed}
}
\description{
see title
}
\keyword{internal}
