% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{HMM_regression}
\alias{HMM_regression}
\alias{HMM_regression.multfsusie}
\title{Compute refined estimate using HMM regression}
\usage{
HMM_regression(multfsusie.obj, Y, X, verbose, maxit, ...)

\method{HMM_regression}{multfsusie}(multfsusie.obj, Y, X, ind_analysis, verbose = TRUE, maxit = 5)
}
\arguments{
\item{multfsusie.obj}{a multfsusie.obj object}

\item{Y}{multfsusie.obj data object}

\item{X}{matrix containing the covariates}

\item{verbose}{logical
@param maxit max number of iteration}
}
\description{
Compute refined estimate using HMM regression
}
