% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO_mutlfsusie.R
\name{loglik_SFR}
\alias{loglik_SFR}
\alias{loglik_SFR.multfsusie}
\title{Compute log likelihood of single function regression of effect l}
\usage{
loglik_SFR(multfsusie.obj, l, Y, X, ...)

\method{loglik_SFR}{multfsusie}(multfsusie.obj, l, Y, X, ind_analysis, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by \code{\link{init_multfsusie_obj}} function}

\item{l}{effect to update}

\item{Y}{wavelet transformed phenotype}

\item{X}{Matrix of covariates}

\item{...}{Additional arguments passed to other functions.}

\item{ind_analysis}{output of which_notNA_pos}
}
\value{
The log-likelihood, \eqn{\log p(Y | X, V)}, where V is the prior parameters
}
\description{
Compute log likelihood of single function regression of effect l
}
\keyword{internal}
