% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{merge_effect}
\alias{merge_effect}
\alias{merge_effect.multfsusie}
\title{Merging effect function}
\usage{
merge_effect(multfsusie.obj, tl, ...)

\method{merge_effect}{multfsusie}(multfsusie.obj, tl, discard = FALSE, ...)
}
\arguments{
\item{multfsusie.obj}{a susiF object defined by init_multfsusie_obj function}

\item{tl}{see  \code{\link{greedy_backfit}}}
}
\value{
a multfusie  object
}
\description{
Merging effect function
}
\keyword{internal}
