% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{out_prep}
\alias{out_prep}
\alias{out_prep.multfsusie}
\title{Preparing output of main multfsusie function}
\usage{
out_prep(
  multfsusie.obj,
  Y,
  interpolated_Y,
  X,
  list_indx_lst,
  filter_cs,
  outing_grid,
  cov_lev,
  filter.number,
  family,
  ind_analysis,
  post_processing,
  posthoc = TRUE,
  posthoc_prior_active = 0.5,
  posthoc_variant_prior = NULL,
  tidx = integer(),
  kept_index = NULL,
  original_P = NULL,
  names_colX = NULL,
  original_mean_X = NULL,
  ...
)

\method{out_prep}{multfsusie}(
  multfsusie.obj,
  Y,
  interpolated_Y,
  X,
  list_indx_lst,
  filter_cs,
  outing_grid,
  cov_lev,
  filter.number = 10,
  family = "DaubLeAsymm",
  ind_analysis,
  post_processing = "smash",
  verbose = TRUE,
  posthoc = TRUE,
  posthoc_prior_active = 0.5,
  posthoc_variant_prior = NULL,
  tidx = integer(),
  kept_index = NULL,
  original_P = NULL,
  names_colX = NULL,
  original_mean_X = NULL,
  ...
)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by init_multfsusie_obj function}

\item{Y}{data}

\item{interpolated_Y}{interpolated functional data}

\item{X}{matrix of size N by p}

\item{list_indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{filter_cs}{logical, if TRUE filter the credible set (removing low purity cs and cs with estimated prior equal to 0)}

\item{post_processing}{the chosen postprocessing}

\item{posthoc}{logical , default TRUE if set to TRUE compute post hoc probabilities of causal configurations as in Yuan Nat Gen 2024}

\item{posthoc_prior_active}{prior probability that an individual modality is
active in a post-hoc configuration.}

\item{posthoc_variant_prior}{optional prior probabilities over covariates for
post-hoc configuration Bayes factors. Defaults to a uniform prior.}

\item{tidx}{original positions of X columns excluded from fitting}

\item{kept_index}{original positions of X columns retained for fitting}

\item{original_P}{number of columns in the user-supplied X}

\item{names_colX}{original column names of X, if present}

\item{original_mean_X}{column means of the user-supplied X}

\item{verbose}{true or false}
}
\description{
Preparing output of main multfsusie function
}
\keyword{internal}
