% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{out_prep}
\alias{out_prep}
\alias{out_prep.multfsusie}
\title{Preparing output of main multfsusie function}
\usage{
out_prep(
  multfsusie.obj,
  Y,
  interpolated_Y,
  X,
  list_indx_lst,
  filter_cs,
  outing_grid,
  cov_lev,
  filter.number,
  family,
  ind_analysis,
  TI,
  HMM,
  ...
)

\method{out_prep}{multfsusie}(
  multfsusie.obj,
  Y,
  interpolated_Y,
  X,
  list_indx_lst,
  filter_cs,
  outing_grid,
  cov_lev,
  filter.number = 10,
  family = "DaubLeAsymm",
  ind_analysis,
  TI = FALSE,
  HMM = FALSE,
  ...
)
}
\arguments{
\item{multfsusie.obj}{a multfsusie object defined by init_multfsusie_obj function}

\item{Y}{data}

\item{interpolated_Y}{interpolated functional data}

\item{X}{matrix of size N by p}

\item{list_indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{filter_cs}{logical, if TRUE filter the credible set (removing low purity cs and cs with estimated prior equal to 0)}
}
\description{
Preparing output of main multfsusie function
}
\keyword{internal}
