% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{out_prep}
\alias{out_prep}
\alias{out_prep.multfsusie}
\title{Preparing output of main multfsusie function}
\usage{
out_prep(mvfsusie.obj, Y, X, indx_lst, filter.cs, lfsr_curve, ...)

\method{out_prep}{multfsusie}(
  multfsusie.obj,
  Y,
  interpolated_Y,
  X,
  list_indx_lst,
  filter.cs,
  outing_grid,
  cov_lev,
  filter.number = 10,
  family = "DaubLeAsymm",
  TI = FALSE
)
}
\arguments{
\item{Y}{data}

\item{X}{matrix of size N by p}

\item{filter.cs}{logical, if TRUE filter the credible set (removing low purity cs and cs with estimated prior equal to 0)}

\item{multfsusie.obj}{a multfsusie object defined by \code{\link{init_multfsusie_obj}} function}

\item{list_indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\description{
Preparing output of main multfsusie function
}
\keyword{internal}
