% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{TI_regression}
\alias{TI_regression}
\alias{TI_regression.multfsusie}
\title{Compute refined estimate using translation invariant wavelet transform}
\usage{
TI_regression(
  multfsusie.obj,
  Y,
  X,
  ind_analysis,
  verbose,
  filter.number,
  family,
  ...
)

\method{TI_regression}{multfsusie}(
  multfsusie.obj,
  Y,
  X,
  ind_analysis,
  verbose = TRUE,
  filter.number = 10,
  family = "DaubLeAsymm",
  ...
)
}
\arguments{
\item{multfsusie.obj}{a multfsusie.obj object}

\item{Y}{multfsusie.obj data object}

\item{X}{matrix containing the covariates}

\item{ind_analysis}{output of which_notNA_pos}

\item{verbose}{logical}

\item{filter.number}{see wd description in wavethesh package description}

\item{family}{see wd description in wavethresh package description}

\item{...}{Additional arguments passed to other functions.}
}
\description{
Compute refined estimate using translation invariant wavelet transform
}
\keyword{internal}
