% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_lBF}
\alias{update_lBF}
\alias{update_lBF.multfsusie}
\title{Update multfsusie log Bayes factor}
\usage{
update_lBF(multfsusie.obj, l, lBF, ...)

\method{update_lBF}{multfsusie}(multfsusie.obj, l, lBF, lBF_per_trait, ...)
}
\arguments{
\item{multfsusie.obj}{a susiF object defined by init_multfsusie_obj function}

\item{l}{effect to update}

\item{lBF}{vector of length p, containing the updated log Bayes factors}
}
\value{
multfsusie object
}
\description{
Update multfsusie log Bayes factor
}
\keyword{internal}
