% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{update_lfsr}
\alias{update_lfsr}
\alias{update_lfsr.multfsusie}
\title{Update lfsr multfsusie l}
\usage{
update_lfsr(multfsusie.obj, l, ...)

\method{update_lfsr}{multfsusie}(multfsusie.obj, l, effect_estimate, list_indx_lst, ...)
}
\arguments{
\item{multfsusie.obj}{a multfsusie}

\item{l}{the effect to be updated new ELBO value}
}
\value{
multfsusie object
}
\description{
Update lfsr multfsusie l
}
\keyword{internal}
