% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_multfsusie_obj.R
\name{which_dummy_cs}
\alias{which_dummy_cs}
\alias{which_dummy_cs.multfsusie}
\title{Return which credible sets are  dummy}
\usage{
which_dummy_cs(multfsusie.obj, min.purity = 0.5, X, ...)

\method{which_dummy_cs}{multfsusie}(multfsusie.obj, min.purity = 0.5, X, median_crit = FALSE)
}
\arguments{
\item{multfsusie.obj}{a susif object defined by \code{\link{init_susiF_obj}} function}

\item{min.purity}{minimal purity within a CS}

\item{X}{matrix of covariate}
}
\value{
a list of index corresponding the the dummy effect
}
\description{
Return which credible sets are  dummy
}
\keyword{internal}
